Refactorización del plugin

+++ Model/ConteoStock.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * This file is part of StockAvanzado plugin for FacturaScripts
- * Copyright (C) 2020-2023 Carlos Garcia Gomez <carlos@facturascripts.com>
+ * Copyright (C) 2020-2025 Carlos Garcia Gomez <carlos@facturascripts.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -24,8 +24,11 @@ use FacturaScripts\Core\DataSrc\Almacenes;
 use FacturaScripts\Core\Model\Base;
 use FacturaScripts\Core\Session;
 use FacturaScripts\Core\Tools;
+use FacturaScripts\Dinamic\Lib\StockMovementManager;
+use FacturaScripts\Dinamic\Lib\StockRebuild;
 use FacturaScripts\Dinamic\Model\Almacen;
-use FacturaScripts\Dinamic\Model\Stock;
+use FacturaScripts\Dinamic\Model\ConteoStock as DinConteoStock;
+use FacturaScripts\Dinamic\Model\LineaConteoStock;
 
 /**
  * Description of ConteoStock
@@ -38,6 +41,9 @@ class ConteoStock extends Base\ModelClass
 
     /** @var string */
     public $codalmacen;
+
+    /** @var bool */
+    public $completed;
 
     /** @var string */
     public $fechafin;
@@ -54,9 +60,42 @@ class ConteoStock extends Base\ModelClass
     /** @var string */
     public $observaciones;
 
+    public function addLine(string $referencia, int $idproducto, float $quantity): LineaConteoStock
+    {
+        $line = new LineaConteoStock();
+        $where = [
+            new DataBaseWhere('idconteo', $this->idconteo),
+            new DataBaseWhere('referencia', $referencia)
+        ];
+        $orderBy = ['idlinea' => 'DESC'];
+
+        // si no existe la línea, la creamos
+        if (false === $line->loadFromCode('', $where, $orderBy)) {
+            $line->cantidad = $quantity;
+            $line->idconteo = $this->idconteo;
+            $line->idproducto = $idproducto;
+            $line->referencia = $referencia;
+        } else {
+            // si ya existe la línea, incrementamos la cantidad
+            $line->cantidad++;
+        }
+
+        $line->fecha = Tools::dateTime();
+        $line->nick = Session::user()->nick;
+
+        $resultLine = $this->pipe('addLine', $line);
+        if (null !== $resultLine) {
+            $line = $resultLine;
+        }
+
+        $line->save();
+        return $line;
+    }
+
     public function clear()
     {
         parent::clear();
+        $this->completed = false;
         $this->fechafin = Tools::date();
         $this->fechainicio = Tools::date();
         $this->nick = Session::user()->nick;
@@ -64,13 +103,53 @@ class ConteoStock extends Base\ModelClass
 
     public function delete(): bool
     {
-        foreach ($this->getLines() as $line) {
+        $conteo = new DinConteoStock();
+        if (false === $conteo->loadFromCode($this->idconteo)) {
+            return false;
+        }
+
+        $newTransaction = false === static::$dataBase->inTransaction() && self::$dataBase->beginTransaction();
+        foreach ($this->getLines(['fecha' => 'DESC']) as $line) {
+            if (false === $this->pipeFalse('deleteLineCounting', $line, $conteo)) {
+                if ($newTransaction) {
+                    self::$dataBase->rollback();
+                }
+                return false;
+            }
+
+            if (false === StockMovementManager::deleteLineCounting($line, $conteo)) {
+                if ($newTransaction) {
+                    self::$dataBase->rollback();
+                }
+                return false;
+            }
+
             if (false === $line->delete()) {
+                if ($newTransaction) {
+                    self::$dataBase->rollback();
+                }
+                return false;
+            }
+
+            if (false === StockRebuild::rebuild($line->idproducto)) {
+                if ($newTransaction) {
+                    self::$dataBase->rollback();
+                }
                 return false;
             }
         }
 
-        return parent::delete();
+        if (false === parent::delete()) {
+            if ($newTransaction) {
+                self::$dataBase->rollback();
+            }
+            return false;
+        }
+
+        if ($newTransaction) {
+            self::$dataBase->commit();
+        }
+        return true;
     }
 
     public function getAlmacen(): Almacen
@@ -78,14 +157,11 @@ class ConteoStock extends Base\ModelClass
         return Almacenes::get($this->codalmacen);
     }
 
-    /**
-     * @return LineaConteoStock[]
-     */
-    public function getLines(): array
+    public function getLines(array $order = []): array
     {
         $lineaConteo = new LineaConteoStock();
         $where = [new DataBaseWhere('idconteo', $this->idconteo)];
-        return $lineaConteo->all($where, ['fecha' => 'DESC'], 0, 0);
+        return $lineaConteo->all($where, $order, 0, 0);
     }
 
     public static function primaryColumn(): string
@@ -101,34 +177,70 @@ class ConteoStock extends Base\ModelClass
     public function test(): bool
     {
         $this->observaciones = Tools::noHtml($this->observaciones);
-
         return parent::test();
     }
 
     public function updateStock(): bool
     {
-        self::$dataBase->beginTransaction();
+        $conteo = new DinConteoStock();
+        if (false === $conteo->loadFromCode($this->idconteo)) {
+            return false;
+        }
 
-        foreach ($this->getLines() as $line) {
-            $stock = new Stock();
-            $where = [
-                new DataBaseWhere('codalmacen', $this->codalmacen),
-                new DataBaseWhere('referencia', $line->referencia)
-            ];
-            if (false === $stock->loadFromCode('', $where)) {
-                // el stock no existe, lo creamos
-                $stock->codalmacen = $this->codalmacen;
-                $stock->referencia = $line->referencia;
+        // si el conteo ya está completado, no hacemos nada
+        if ($conteo->completed) {
+            return true;
+        }
+
+        // establecemos la fecha de fin del conteo
+        $conteo->fechafin = Tools::dateTime();
+
+        // primero recorremos las líneas para obtener el stock actual por referencia
+        $stocks = [];
+        foreach ($conteo->getLines() as $line) {
+            if (false === isset($stocks[$line->referencia])) {
+                $stocks[$line->referencia] = $line->cantidad;
+                continue;
+            }
+            $stocks[$line->referencia] += $line->cantidad;
+        }
+
+        $newTransaction = false === static::$dataBase->inTransaction() && self::$dataBase->beginTransaction();
+        foreach ($conteo->getLines(['fecha' => 'ASC']) as $line) {
+            if (false === StockMovementManager::addLineCounting($line, $conteo, $stocks[$line->referencia])) {
+                if ($newTransaction) {
+                    self::$dataBase->rollback();
+                }
+                return false;
             }
 
-            $stock->cantidad = $line->cantidad;
-            if (false === $stock->save()) {
-                self::$dataBase->rollback();
+            if (false === $this->pipeFalse('updateStock', $line, $conteo, $stocks[$line->referencia])) {
+                if ($newTransaction) {
+                    self::$dataBase->rollback();
+                }
+                return false;
+            }
+
+            if (false === StockRebuild::rebuild($line->idproducto)) {
+                if ($newTransaction) {
+                    self::$dataBase->rollback();
+                }
                 return false;
             }
         }
 
-        self::$dataBase->commit();
+        //actualizamos el conteo
+        $conteo->completed = true;
+        if (false === $conteo->save()) {
+            if ($newTransaction) {
+                self::$dataBase->rollback();
+            }
+            return false;
+        }
+
+        if ($newTransaction) {
+            self::$dataBase->commit();
+        }
         return true;
     }
